Planet models, sliders and etc. updated
Added models for all the planets of the solra system.
Currently the moon is not implemented.
Slider max value adjusted from 3 to 7, to be able to choose all the planets of the solar system ( pluto excluded ).
Added correctly scaled sizes+mass of planets.
Might need to change how planets orbit eachother...