Star parallax/shine color fix
Stars parallax slightly when camera moves, since the distance between the orbiting planets and the stars is so massive, a small parallax has been added.
Small stars flash due to size constraints, but it is intentional to add depth to the background.
Colors now correctly pass through to the parallax shader, the shader made all stars white before, fixed.